.vscode added to gitignore, Test controllers updated to non-deprecated class hierarchy
